backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
websockets==12.0

# Database
prisma==0.11.0
psycopg2-binary==2.9.9
redis==5.0.1

# AI & ML
openai==1.12.0
sentence-transformers==2.3.1
faiss-cpu==1.8.0
xgboost==2.0.3
scikit-learn==1.4.0
numpy==1.26.3
pandas==2.2.0
cryptography==42.0.0

# Security & Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1
httpx==0.26.0

# Notifications
twilio==8.11.1
requests==2.31.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
